+++ src/stactools/testing/cli_test.py
@@ -1,14 +1,17 @@
 import logging
 import unittest
 from abc import ABC, abstractmethod
-from typing import Callable, List
+from typing import Callable, List, Optional, Sequence, Union
 
 import click
 from click.testing import CliRunner, Result
 
 
 class CliTestCase(unittest.TestCase, ABC):
+    """A command-line interface test case."""
+
     def use_debug_logging(self) -> None:
+        """Enable debug logging for these tests."""
         logger = logging.getLogger("stactools")
         logger.setLevel(logging.DEBUG)
 
@@ -20,6 +23,8 @@ class CliTestCase(unittest.TestCase, ABC):
         logger.addHandler(ch)
 
     def setUp(self) -> None:
+        """Sets up a mock cli group for testing."""
+
         @click.group()
         def cli() -> None:
             pass
@@ -28,7 +33,17 @@ class CliTestCase(unittest.TestCase, ABC):
             create_subcommand(cli)
         self.cli = cli
 
-    def run_command(self, cmd: str) -> Result:
+    def run_command(self, cmd: Optional[Union[str, Sequence[str]]]) -> Result:
+        """Runs a command, returning its result.
+
+        If there is output, print it to stdout.
+
+        Args:
+            cmd (str): The command to run.
+
+        Returns:
+            click.Result: The command-line invocation result.
+        """
         runner = CliRunner()
         result = runner.invoke(self.cli, cmd, catch_exceptions=False)
         if result.output:
@@ -39,5 +54,9 @@ class CliTestCase(unittest.TestCase, ABC):
     def create_subcommand_functions(
         self,
     ) -> List[Callable[[click.Group], click.Command]]:
-        """Return list of 'create_command' functions from implementations"""
+        """Return list of 'create_command' functions from implementations.
+
+        Returns:
+            list[Callable[[click.Group], click.Command]]: The commands to run.
+        """
         pass