+++ CHANGELOG.md
@@ -9,8 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added geometry.py for handling common geometrical operations like bounding boxes from GeoJSON ([#314](https://github.com/stac-utils/stactools/pull/314))
 - Specify installation channel to use for all conda packages to avoid incompatibility ([#301](https://github.com/stac-utils/stactools/pull/301))
+- Allow MultiPolygons when fixing antimeridian issues ([#317](https://github.com/stac-utils/stactools/pull/317))
 - Added `raster_footprint` module to assist in populating the geometry of an Item from data coverage of its data assets ([#307](https://github.com/stac-utils/stactools/pull/307))
+
+### Changed
+
+- Modified stactools.core.utils.antimeridian.fix_item to return the item and updated 2 unit tests ([#309](https://github.com/stac-utils/stactools/issues/309))
+- Relaxed typing for cmd parameter for the CliTestCase.run_command in cli_test.py ([#306](https://github.com/stac-utils/stactools/issues/306))
+- Cleaned up API documentation ([#315](https://github.com/stac-utils/stactools/pull/315))
+
+### Removed
+
+- Unnecessary and incorrect `args` and `kwargs` from `StacIO` subclass ([#315](https://github.com/stac-utils/stactools/pull/315))
+
+### Removed
+
+- Dropped support for Python 3.7 ([#223](https://github.com/stac-utils/stactools/issues/223))
 
 ## [v0.3.1]